Decouple the normalized Event model from Ticketmaster's vendor shape
Rescoped Phase 1 of the roadmap ("Foundation — normalized Event
model") after a grilling session: the pipeline-unification work it
was originally scoped against already happened (#40, "Collapse three
Ticketmaster event fetchers into one module"). What's left, and what
ships here, is decoupling — not building a multi-source adapter
against a PredictHQ shape nobody's validated yet (that stays deferred
until Phase 2's own spike).

- New apps/api/src/events/ module: EventResponse and its constituent
  types (renamed off their Tm-prefixed originals: Performer, PriceRange,
  Classification/Segment, ExternalLinks/ExternalLink, plus Images/
  VenueResponse/LocationResponse), and the source-agnostic logic that
  operates on them (apply_personalization, dedupe_key). Ticketmaster's
  raw shapes (TmEvent, TmDate*, EventEmbedded, TmVenue/City/Location)
  and normalize_event — the one function that still needs to know
  Ticketmaster's shape — stay in ticketmaster_stream.
- EventResponse.attractions -> performers, Rust field and JSON key both,
  cascading through every consumer (playlist_builder.rs on the backend;
  EventCard, EventFilter, EventDetails, reducers/events.ts, and
  eventsStream.ts on the frontend).
- Collapsed the frontend's duplicate lib/types.ts TmEvent/Event type
  (already-drifted, missing matchedArtist/matchedVia) into the real
  EventResponse from eventsStream.ts.
