+++ apps/api/src/ticketmaster_stream/types.rs
@@ -1,7 +1,9 @@
-//! Raw Ticketmaster API response shapes, and the normalized shapes this
-//! service returns to clients.
+//! Raw Ticketmaster API response shapes, and the query/control types this
+//! module's handlers accept. The normalized response shape lives in
+//! `crate::events::types` — see that module for why.
 
-use serde::{Deserialize, Serialize};
+use crate::events::types::{Classification, Images, Performer, PriceRange};
+use serde::Deserialize;
 
 #[derive(Deserialize)]
 pub struct EventsQuery {
@@ -56,21 +58,14 @@ pub struct TmEvent {
     pub id: String,
     pub url: Option<String>,
     #[serde(rename = "priceRanges")]
-    pub price_ranges: Option<Vec<TmPriceRange>>,
+    pub price_ranges: Option<Vec<PriceRange>>,
     pub name: String,
     #[serde(default)]
     pub images: Vec<Images>,
     pub dates: TmDate,
-    pub classifications: Option<Vec<TmClassification>>,
+    pub classifications: Option<Vec<Classification>>,
     #[serde(rename = "_embedded")]
     pub embedded: Option<EventEmbedded>,
-}
-
-#[derive(Debug, Deserialize, Serialize, Clone)]
-pub struct TmPriceRange {
-    currency: String,
-    min: f32,
-    max: f32,
 }
 
 #[derive(Debug, Deserialize)]
@@ -91,47 +86,7 @@ pub struct TmDateStart {
 #[derive(Debug, Deserialize)]
 pub struct EventEmbedded {
     pub venues: Option<Vec<TmVenue>>,
-    pub attractions: Option<Vec<TmAttraction>>,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct TmAttraction {
-    pub name: Option<String>,
-    pub id: Option<String>,
-    pub classifications: Option<Vec<TmClassification>>,
-    #[serde(rename = "externalLinks")]
-    pub external_links: Option<TmExternalLinks>,
-    pub images: Option<Vec<Images>>,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct TmExternalLinks {
-    wiki: Option<Vec<TmExternalLink>>,
-    homepage: Option<Vec<TmExternalLink>>,
-    instagram: Option<Vec<TmExternalLink>>,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct TmExternalLink {
-    url: Option<String>,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct TmClassification {
-    primary: Option<bool>,
-    segment: Option<TmSegment>,
-    genre: Option<TmSegment>,
-    #[serde(rename = "subGenre")]
-    sub_genre: Option<TmSegment>,
-    #[serde(rename = "subType")]
-    sub_type: Option<TmSegment>,
-    family: Option<bool>,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct TmSegment {
-    id: String,
-    name: String,
+    pub attractions: Option<Vec<Performer>>,
 }
 
 #[derive(Debug, Deserialize, Clone)]
@@ -148,56 +103,6 @@ pub struct TmCity {
 
 #[derive(Debug, Deserialize, Clone)]
 pub struct TmLocation {
-    pub latitude: Option<String>,
-    pub longitude: Option<String>,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct Images {
-    ratio: Option<String>,
-    url: String,
-    width: Option<i32>,
-    height: Option<i32>,
-    fallback: Option<bool>,
-}
-
-#[derive(Debug, Serialize, Clone)]
-pub struct EventResponse {
-    pub id: String,
-    pub name: String,
-    pub venue: Option<VenueResponse>,
-    pub images: Vec<Images>,
-    pub dates: Option<String>,
-    #[serde(rename = "datesPretty")]
-    pub dates_pretty: Option<String>,
-    pub classifications: Option<Vec<TmClassification>>,
-    pub attractions: Option<Vec<TmAttraction>>,
-    pub url: Option<String>,
-    #[serde(rename = "priceRanges")]
-    pub price_ranges: Option<Vec<TmPriceRange>>,
-    /// Name of the attraction that matched one of the caller's Spotify top
-    /// artists, for personalized discovery. `None` when the caller isn't
-    /// connected, has no personalization data, or no attraction matched.
-    #[serde(rename = "matchedArtist")]
-    pub matched_artist: Option<String>,
-    /// Set only when `matched_artist` matched via Apple Music similar-artist
-    /// expansion rather than being one of the caller's own Spotify top
-    /// artists — names the top artist (`seed`) that produced the match, so
-    /// the frontend can render "similar to X" instead of implying the
-    /// caller directly listens to `matched_artist`.
-    #[serde(rename = "matchedVia")]
-    pub matched_via: Option<String>,
-}
-
-#[derive(Debug, Serialize, Clone)]
-pub struct VenueResponse {
-    pub name: Option<String>,
-    pub location: Option<LocationResponse>,
-    pub city: Option<String>,
-}
-
-#[derive(Debug, Serialize, Clone)]
-pub struct LocationResponse {
     pub latitude: Option<String>,
     pub longitude: Option<String>,
 }
